First Commit
Changed the settings for how Visual Studio interacts with Git. Included a set of files to be ignored while pushing something to the cloud repository.

"I had the same problem on the Team Explorer-Home go to settings -> repository settings-> Ignore & Attributes File -> click on add then on Team Explorer-Home-> global settings -> enable push -- force and that worked for me ."